pull test code into test-log.cpp; add to Makefile
